Move some parsing functions out, start to work on moving things around
