+++ src/Components/Cover.js
@@ -3,7 +3,7 @@ import { FaFilm, FaCheck } from 'react-icons/fa';
 import levenshtein from 'js-levenshtein';
 import './Cover.css';
 // import Spinner from './Spinner';
-import ScrollReveal from '../ScrollReveal';
+import ScrollReveal from '../Util/ScrollReveal';
 
 class Cover extends Component {
     componentDidMount() {
